reorg network context options.
